pyproject.toml
[tool.poetry]
name = "qraft"
version = "0.1.0"
description = "Qraft7.0 - 数据流驱动自动化模式探索平台"
authors = ["Qraft Team <team@qraft.example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asyncio = "*"
uvloop = "^0.17.0"
polars = "^0.19.0"
pydantic = "^2.4.0"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
nats-py = "^2.3.1"
clickhouse-driver = "^0.2.6"
asyncx = "^0.0.3"
river = "^0.15.0"
scikit-learn = "^1.3.0"
prometheus-client = "^0.17.0"
pyyaml = "^6.0.1"
loguru = "^0.7.0"
uuid = "^1.30"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"